Remove invalid CSS value; remove unnecessary non-themed styling.
`list-style-type:bullet` is invalid. The correct value is `disc`. As of 2016/6/2, Firefox incorrectly renders `<ul>` elements with `list-style-type:bullet` as if they had `list-style-type: decimal`, even though the element is a `<ul>` instead of a `<ol>`. We believe this is a bug in Firefox, triggered by a bug in our CSS. We are eliminating the bug from our CSS for now.
The browser default values for the other properties dropped in this commit are acceptable for a stylesheet that doesn't implement a theme. Explicit rendering of lists are better handled by themed stylesheets, instead.