Add a `.highlights` container to `.mainText`
In support of the highlights we will feature on a Parks activity page, add a container that displays as a right sidebar on a medium-sized viewport or higher.

Also, `.mainText` doesn't have a `<header>`,  so it doesn't have a consistent model with `.uiBlock`. This commit includes some changes that extricate `.mainText` from `.uiBlock`.