`.mainText` -> `.contentBlock`
The `.mainText` element doesn't have the same model as a `.uiBlock`, but serves a critical function, so it made sense to give it a name that better expresses its purpose.

It's a block, but it's not a block for displaying UI. Its purpose is to display content, usually stored by a CMS. It has a specific `div` for styling user content, and permits neighboring elements (such as `.highlight`) to help enhance content for end users.