Adjustments to width of tabular data
A table is currently treated as a 12-column grid element, because it is a child of a grid container. However, `width: auto` doesn't yield the same layout behavior with a `<table>` as it does with a `<div>`. It is undesireable to set `width: 100%` because that would make the table expand outside of the gutters of the grid system.

Instead of setting width at 100%, this commit manually sets a fixed width for tables when the viewport is wider than the maximum layout width, and provides an eyeball estimated proportional width when the viewport is larger than a "large" viewport, but less than the maximum layout width.